Add support for up/down navigation.
This turned out to be the most complex PR I've made so far. In order to
make this work, I mirror the text of the current note row up to the
cursor to an off-screen mirror, which also auto-resizes. I then measure
the height of that and divide by a fixed pixel value, which roughly
corresponds to the height of a row. This is disgustingly fragile but
seems to work right now.